+++ app/src/main/java/ru/fefu/pnexpert/MainActivity.kt
@@ -3,23 +3,24 @@ package ru.fefu.pnexpert
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import ru.fefu.pnexpert.ui.theme.PNExpertTheme
+import ru.fefu.pnexpert.ui.theme.PnExpertTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            PNExpertTheme {
+            PnExpertTheme() {
                 // A surface container using the 'background' color from the theme
 
                 Surface(
                     modifier = Modifier
                         .fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
+                    color = PnExpertTheme.colors.AppBlueLightColor
                 ) {
 
                 }
