+++ app/src/main/java/ru/fefu/pnexpert/ui/theme/PnExpertColor.kt
@@ -1,0 +1,57 @@
+package ru.fefu.pnexpert.ui.theme
+
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+
+val baseAppPalette = PnExpertColors(
+    //MainAppColors
+    AppDarkColor= Color(0xFF393939),
+    AppBlueColor= Color(0xFF5668E8),
+    AppPinkLightColors = Color(0xFFF9C9DC),
+    AppGreyLightColor= Color(0xFFF4F4F4),
+    AppGreyDarkColor= Color(0xFF8B8B8B),
+    AppBlueLightColor= Color(0xFF6AEBEB),
+    AppWhiteColor = Color(0xFFFFFFFF),
+    AppPinkDarkColor = Color(0xFFF25F8E),
+
+    //ButtonsColors
+    ButtonNormalBlueColor = Color(0xFF5668E8),
+    ButtonNormalRedColor = Color(0xFFF25F8E),
+
+    ButtonHoverBlueColor = Color(0xFF3E50CE),
+    ButtonHoverRedColor = Color(0xFFEB4E80),
+    ButtonHoverLightColor = Color(0xFFE5E9FF),
+
+    ButtonPressedBlueColor = Color(0xFF2C3CAC),
+    ButtonPressedRedColor= Color(0xFFD6376A),
+    ButtonPressedLightColor = Color(0xFFC7CEFF),
+
+    ButtonInactiveColor = Color(0xFF929FFF),
+
+    //FontColors
+    FontDarkColor= Color(0xFF393939),
+    FontBlueColor = Color(0xFF5668E8),
+    FontGreyColor = Color(0xFF8B8B8B),
+    FontWhiteColor = Color(0xFFFFFFFF),
+
+    //IconColors
+    IconBlueColor = Color(0xFF5668E8),
+    IconBlueLightColor = Color(0xFF5FABF2),
+    IconRedColor = Color(0xFFF24726),
+    IconPinkColor = Color(0xFFF25F8E),
+    IconGreenColor = Color(0xFF149D17),
+    IconOrangeColor = Color(0xFFFA8E10),
+
+    //Gradients
+    GradientPink = Brush.linearGradient(
+        listOf(Color(0xFF5668E8), Color(0xFFF9C9DC)),
+        start = Offset(0.5f, 0f),
+        end = Offset(0.5f, 1f)
+    ),
+    GradientBlue = Brush.linearGradient(
+        listOf(Color(0xFF5668E8), Color(0xFF6AEBEB)),
+        start = Offset(0f, 0.5f),
+        end = Offset(1f, 0.5f)
+    )
+)
